+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - "Exchange Management Shell" mode support.
 - Time filters.
+### Changed
+- Make Script more modular to enable adding more tools.
 
 ## [0.0.1] - 2020-12-28
 ### Added